Multiple improvements
Improved handling of "$_flag_f" force fetch. Note the errors I had been
seeing earlier when trying to pass the flag and re-parse it were totally
unrelated to the flag and parsing. I was using echo to debug and that
echo would of course later  on be redirected into a function ~_~ where
it caused an error.

Improved error messages. Now checks if there's a yarn.js in /bin. Also
prints more info on failing curl.

Improved help command with info about supported versions.

It now uses the tarball URL from the release to fetch the tarball which
should be more resilient.