Only fetch release if needed
Check if version user wants to activate is already installed and only
get releases if that's not the case. This fixes the issue that when
switching from one version that's installed to another that's installed,
releases are fetched from origin if they're older than N.

latest is exempt from this since latest can have changed in the
meantime.