Improve error handling from curl
The redirection of _yvm_get_releases into a variable turned out to be a
problem. I don't know how to exit early from that, since I can't use
return in a pipeline and exit will also quit the shell. I also can't use
return in the (_yvm_get_releases || return 1) calls.

The workaround for now is to save the curl call in a temp file. Now I
can do the usual || return 1 pattern. I also replaced -s with
--no-progress-meter so users see curl errors.

The end result of this is that an error from curl will stop the program
and the error should be visible. Same for errors from the tr awk
pipeline.